Move DateTimeStringValidator to helper module
