+++ linode/helper/framework_validators.go
@@ -1,4 +1,4 @@
-package token
+package helper
 
 import (
 	"context"
